Fixed some warnings

+++ OAuthStarterKit/OAMutableURLRequest.m
@@ -138,7 +138,7 @@ signatureProvider:(id<OASignatureProviding, NSObject>)aProvider
 - (void)_generateTimestamp 
 {
 	[timestamp release];
-    timestamp = [[NSString alloc]initWithFormat:@"%d", time(NULL)];
+    timestamp = [[NSString alloc]initWithFormat:@"%d", (int) time(NULL)];
 }
 
 - (void)_generateNonce
